+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ### 0.1.10
+* Fix `Naut` not updating tokens when base URI is updated
 
 ### 0.1.9
 * Implement `Naut` module