+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 0.1.10
 * Fix `Naut` not updating tokens when base URI is updated
+* Fix "could not find ABI for contract `IERC721` in `naut`"
 
 ### 0.1.9
 * Implement `Naut` module