+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 0.1.10
 * Fix `Naut` not updating tokens when base URI is updated
 * Implement `Planet` module
+* Implement `Delegation` module
 
 ### 0.1.9
 * Implement `Naut` module